Split other methods in txn into separate files
